Split room info/settings into protocol-specific components
Also show a loading spinner while room info hasn't been fetched yet
