+++ app/components/channel-settings-irc.hbs
@@ -1,0 +1,4 @@
+<div class="bg-white bg-opacity-90 p-4 rounded border border-white/20 text-neutral-800 shadow-sm">
+  <h3 class="text-lg font-bold mb-1 text-neutral-900">{{@channel.displayName}}</h3>
+  <p class="text-sm text-neutral-500">{{@channel.name}}</p>
+</div>
