+++ app/components/loading-spinner.hbs
@@ -1,3 +1,3 @@
-<span class={{this.classNames}} role="status">
+<span class={{this.classNames}} role="status" ...attributes>
   <span class="sr-only">Loading</span>
 </span>
