+++ app/models/base_channel.js
@@ -18,6 +18,7 @@ export default class BaseChannel {
   @tracked unreadMessages = false;
   @tracked unreadMentions = false;
   @tracked visible = false; // Current/active channel
+  @tracked roomInfoLoaded = false;
   @tracked description = null;
   @tracked roomFeatures = new TrackedArray([]);
   @tracked roomIdentities = new TrackedArray([]);
