+++ app/templates/channel/settings.hbs
@@ -8,58 +8,10 @@
   </div>
 
   <div class="px-4 space-y-4">
-    <div class="bg-white bg-opacity-90 p-4 rounded border border-white/20 text-neutral-800 shadow-sm">
-      <h3 class="text-lg font-bold mb-1 text-neutral-900">{{this.channel.model.displayName}}</h3>
-      <p class="text-sm text-neutral-500 mb-4">{{this.channel.model.name}}</p>
-
-      {{#if this.channel.model.description}}
-        <div class="mb-4">
-          <h4 class="text-xs uppercase font-semibold text-neutral-400 tracking-wider mb-1">Description</h4>
-          <p class="text-neutral-800">{{this.channel.model.description}}</p>
-        </div>
-      {{/if}}
-
-      {{#if this.channel.model.roomInfoData.lang.value}}
-        <div class="mb-4">
-          <h4 class="text-xs uppercase font-semibold text-neutral-400 tracking-wider mb-1">Language</h4>
-          <p class="text-neutral-800">{{this.channel.model.roomInfoData.lang.value}}</p>
-        </div>
-      {{/if}}
-    </div>
-
-    {{! Room Configurations Card }}
-    {{#if this.filteredRoomConfigData}}
-      <div class="bg-white bg-opacity-90 p-4 rounded border border-white/20 text-neutral-800 shadow-sm">
-        <h3 class="text-xs font-semibold uppercase tracking-wider text-neutral-400 mb-3">Room Configurations</h3>
-        <dl class="grid grid-cols-1 gap-x-4 gap-y-4 sm:grid-cols-2">
-          {{#each-in this.filteredRoomConfigData as |key field|}}
-            {{#if field.label}}
-              <div>
-                <dt class="text-xs font-medium text-neutral-500">{{field.label}}</dt>
-                <dd class="mt-1 text-neutral-900 font-semibold">
-                  {{#if (eq field.type "boolean")}}
-                    {{if field.value "Yes" "No"}}
-                  {{else}}
-                    {{field.value}}
-                  {{/if}}
-                </dd>
-              </div>
-            {{/if}}
-          {{/each-in}}
-        </dl>
-      </div>
-    {{/if}}
-
-    {{! Supported Features Card }}
-    {{#if this.channel.model.roomFeatures}}
-      <div class="bg-white bg-opacity-90 p-4 rounded border border-white/20 text-neutral-800 shadow-sm">
-        <h3 class="text-xs font-semibold uppercase tracking-wider text-neutral-400 mb-3">Supported Features</h3>
-        <ul class="list-disc pl-5 text-sm text-neutral-800 space-y-1">
-          {{#each this.channel.model.roomFeatures as |feature|}}
-            <li><code>{{feature}}</code></li>
-          {{/each}}
-        </ul>
-      </div>
+    {{#if (eq this.channel.model.protocol "XMPP")}}
+      <ChannelSettingsXmpp @channel={{this.channel.model}} />
+    {{else if (eq this.channel.model.protocol "IRC")}}
+      <ChannelSettingsIrc @channel={{this.channel.model}} />
     {{/if}}
   </div>
 </section>
